chore: - Refactored Storybook configuration to use absolute paths for addons and builder. - Removed badges configuration and related imports from preview.js. - Updated theme import in theme.js to use the new path. - Upgraded Storybook and related addons to version 9.1.0-alpha.1. - Fixed repository URL in proprietary/assets package.json.
