+++ components/paragraph/paragraph.stories.js
@@ -3,7 +3,6 @@ import leadDocs from './docs/_lead.md?raw';
 import readme from './README.md?raw';
 import '@dictu/design-tokens/dist/paragraph.css';
 import './dist/index.css';
-import { BADGES } from '@dictu/storybook/config/preview';
 
 export default {
   args: {
@@ -19,7 +18,6 @@ export default {
     },
   },
   parameters: {
-    badges: [BADGES.WIP],
     docs: {
       description: {
         component: readme,
