+++ packages/storybook/config/manager.js
@@ -1,4 +1,4 @@
-import { addons } from '@storybook/manager-api';
+import { addons } from 'storybook/manager-api';
 import theme from './theme';
 
 addons.setConfig({
