+++ packages/storybook/package.json
@@ -23,21 +23,15 @@
   "devDependencies": {
     "@babel/core": "7.26.10",
     "@etchteam/storybook-addon-status": "5.0.0",
-    "@geometricpanda/storybook-addon-badges": "2.0.5",
-    "@storybook/addon-a11y": "8.6.12",
-    "@storybook/addon-actions": "8.6.12",
-    "@storybook/addon-controls": "8.6.12",
-    "@storybook/addon-docs": "8.6.12",
-    "@storybook/addon-interactions": "8.6.12",
-    "@storybook/addon-links": "8.6.12",
-    "@storybook/addon-viewport": "8.6.12",
-    "@storybook/html": "8.6.12",
-    "@storybook/html-vite": "8.6.12",
-    "@storybook/manager-api": "8.6.12",
-    "@storybook/theming": "8.6.12",
-    "@whitespace/storybook-addon-html": "^7.0.0",
+    "@storybook/addon-a11y": "9.1.0-alpha.1",
+    "@storybook/addon-docs": "9.1.0-alpha.1",
+    "@storybook/addon-links": "9.1.0-alpha.1",
+    "@storybook/html-vite": "9.1.0-alpha.1",
     "rimraf": "6.0.1",
-    "storybook": "8.6.12",
+    "storybook": "9.1.0-alpha.1",
     "vite": "^6.3.5"
+  },
+  "overrides": {
+    "storybook": "$storybook"
   }
 }
